refactor: consolidate content script into MarkerApp architecture
